requirements.txt
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.12
openai>=1.0.0
faiss-cpu>=1.7.4
pypdf2>=3.0.1
python-dotenv>=1.0.0
streamlit>=1.28.0
duckduckgo-search>=3.9.0
chromadb>=0.4.0
tiktoken>=0.5.0
fastapi
uvicorn